fix(dds): 修复 rtps_message.c 编译错误 — MISRA 工具级联损坏(对强制转换表达式赋值)
rtps_message_estimate_size() 中 '(unsigned int)((uint32_t)(size)) += ...' 是
MISRA 自动修复工具产生的非法赋值（不能对 cast 表达式赋值），导致
eth_dds_middleware 编译失败，进而阻塞依赖它的 eth_bootloader 构建。
恢复为 'size += RTPS_SUBMESSAGE_HEADER_SIZE;'（宏为 4U，两侧均为无符号，
不违反 MISRA C:2023 10.4）。

+++ src/dds/rtps/rtps_message.c
@@ -1004,7 +1004,7 @@ bool rtps_message_validate(const uint8_t *data, uint32_t len)
 uint32_t rtps_message_estimate_size(uint32_t data_len, bool has_inline_qos)
 {
     uint32_t size = RTPS_HEADER_SIZE;
-    (unsigned int)((uint32_t)(size)) += RTPS_SUBMESSAGE_HEADER_SIZE;
+    size += RTPS_SUBMESSAGE_HEADER_SIZE;
     size += 20U;  /* DATA submessage fixed fields */
     if (has_inline_qos) {
         size += 8U;  /* 预估QoS字段 */
